move card history, subscriber to card module, fix history
