+++ alcs-api/src/card/card-history/card-history.entity.ts
@@ -1,6 +1,6 @@
 import { Column, Entity, ManyToOne } from 'typeorm';
-import { Card } from '../card/card.entity';
-import { EntityHistory } from '../common/entities/history.entity';
+import { Card } from '../card.entity';
+import { EntityHistory } from '../../common/entities/history.entity';
 
 @Entity()
 export class CardHistory extends EntityHistory {
