+++ alcs-api/src/card/card.module.ts
@@ -1,5 +1,7 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
+import { CardHistory } from './card-history/card-history.entity';
+import { CardSubscriber } from './card-history/card.subscriber';
 import { CardStatusController } from './card-status/card-status.controller';
 import { CardStatus } from './card-status/card-status.entity';
 import { CardStatusService } from './card-status/card-status.service';
@@ -18,10 +20,16 @@ import { CardService } from './card.service';
       CardType,
       CardSubtaskType,
       CardSubtask,
+      CardHistory,
     ]),
   ],
   controllers: [CardStatusController],
-  providers: [CardStatusService, CardService, CardSubtaskService],
+  providers: [
+    CardStatusService,
+    CardService,
+    CardSubtaskService,
+    CardSubscriber,
+  ],
   exports: [CardStatusService, CardService, CardSubtaskService],
 })
 export class CardModule {}
